+++ apps/recommendation_v2/.env.template
@@ -9,6 +9,7 @@ SQL_BASE_USER=
 SQL_BASE_PASSWORD=
 
 # Google Cloud Platform & Vertex AI Services
+ENV_SHORT_NAME=
 GCP_PROJECT=
 VERTEX_RETRIEVAL_ENDPOINT_NAME=
 VERTEX_RANKING_ENDPOINT_NAME=
